fixed authorization for all endpoints

+++ src/main/java/com/jkh98/qna/controller/AnswerController.java
@@ -1,7 +1,9 @@
 package com.jkh98.qna.controller;
 
+import com.jkh98.qna.exception.ForbiddenException;
 import com.jkh98.qna.exception.ResourceNotFoundException;
 import com.jkh98.qna.model.Answer;
+import com.jkh98.qna.model.User;
 import com.jkh98.qna.repository.AnswerRepository;
 import com.jkh98.qna.repository.QuestionRepository;
 import com.jkh98.qna.repository.UserRepository;
@@ -9,6 +11,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -36,8 +40,14 @@ public class AnswerController {
     @PostMapping("/questions/{questionId}/answers")
     public Answer addAnswer(@PathVariable UUID questionId,
                             @Valid @RequestBody Answer answer) {
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication()
+                .getPrincipal();
+        User user = userRepository.findByUsername(userDetails.getUsername())
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with username " + userDetails.getUsername()));
+
         return questionRepository.findById(questionId)
                 .map(question -> {
+                    answer.setUser(user);
                     answer.setQuestion(question);
                     return answerRepository.save(answer);
                 }).orElseThrow(() -> new ResourceNotFoundException("Question not found with id " + questionId));
@@ -69,12 +79,21 @@ public class AnswerController {
     public Answer updateAnswer(@PathVariable UUID questionId,
                                @PathVariable UUID answerId,
                                @Valid @RequestBody Answer answerRequest) {
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication()
+                .getPrincipal();
+        User user = userRepository.findByUsername(userDetails.getUsername())
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with username " + userDetails.getUsername()));
+
         if (!questionRepository.existsById(questionId)) {
             throw new ResourceNotFoundException("Question not found with id " + questionId);
         }
 
         return answerRepository.findById(answerId)
                 .map(answer -> {
+                    if (!answer.getUser().equals(user)) {
+                        throw new ForbiddenException("Cannot edit answer");
+                    }
+
                     answer.setText(answerRequest.getText());
                     return answerRepository.save(answer);
                 }).orElseThrow(() -> new ResourceNotFoundException("Answer not found with id " + answerId));
@@ -83,15 +102,24 @@ public class AnswerController {
     @DeleteMapping("/questions/{questionId}/answers/{answerId}")
     public ResponseEntity<?> deleteAnswer(@PathVariable UUID questionId,
                                           @PathVariable UUID answerId) {
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication()
+                .getPrincipal();
+
+        User user = userRepository.findByUsername(userDetails.getUsername())
+                .orElseThrow(() -> new ResourceNotFoundException("User not found with username " + userDetails.getUsername()));
+
         if (!questionRepository.existsById(questionId)) {
             throw new ResourceNotFoundException("Question not found with id " + questionId);
         }
 
         return answerRepository.findById(answerId)
                 .map(answer -> {
+                    if (!answer.getUser().equals(user)) {
+                        throw new ForbiddenException("Cannot delete answer");
+                    }
+
                     answerRepository.delete(answer);
                     return ResponseEntity.ok().build();
                 }).orElseThrow(() -> new ResourceNotFoundException("Answer not found with id " + answerId));
-
     }
 }
